fixed a jq error when selecting the export name for the EC2ASGName on the capacity providers for EC2 page.
The error occurs when you select an array and e.g. one of the items misses an attribute while others have it. In this case one output did not provide an ExportName and the jq select query failed. Wit the fix only entries that do have the ExportName set are selected.